Added error logic in methods

+++ services/fighterService.js
@@ -3,17 +3,21 @@ const { FighterRepository } = require('../repositories/fighterRepository');
 class FighterService {
     // TODO: Implement methods to work with fighters
     getAll() {
-        const item = FighterRepository.getAll();
-        if (!item) {
-            return null;
+        const items = FighterRepository.getAll();
+        if (items.length === 0) {
+            const err = new Error('Fighters not found');
+            err.code = 404;
+            throw err;
         }
-        return item;
+        return items;
     }
     
     search(search) {
         const item = FighterRepository.getOne(search);
         if (!item) {
-            return null;
+            const err = new Error('Fighter not found');
+            err.code = 404;
+            throw err;
         }
         return item;
     }
@@ -21,25 +25,35 @@ class FighterService {
     create(data) {
         const item = FighterRepository.create(data);
         if(!item) {
-            return null;
+            const err = new Error('Fighter not created');
+            err.code = 400;
+            throw err;
         }
         return item;
     }
 
     update(id, data) {
-        const item = FighterRepository.update(id, data);
-        if (!item) {
-            return null;
+        if (this.search({ id })) {
+            const updatedItem = FighterRepository.update(id, data);
+            if (!updatedItem) {
+                const err = new Error('Fighter not updated');
+                err.code = 400;
+                throw err;
+            }
+            return updatedItem;
         }
-        return item;
     }
     
     delete(id) {
-        const item = FighterRepository.delete(id);
-        if (!item) {
-            return null;
+        if (this.search({ id })) {
+            const deletedItem = FighterRepository.delete(id);
+            if (!deletedItem) {
+                const err = new Error('Fighter not deleted');
+                err.code = 400;
+                throw err;
+            }
+            return deletedItem;
         }
-        return item;
     }
 
 }
